fix: Fix build for skia
