+++ src/Uno.UI/UI/Xaml/Controls/ScrollContentPresenter/ScrollContentPresenter.cs
@@ -8,8 +8,7 @@ using System.Collections.Generic;
 using Uno.Disposables;
 using System.Runtime.CompilerServices;
 using System.Text;
-using System.Drawing;
-
+using Windows.Foundation;
 #if XAMARIN_ANDROID
 using View = Android.Views.View;
 using Font = Android.Graphics.Typeface;
@@ -61,11 +60,7 @@ namespace Windows.UI.Xaml.Controls
 				return false;
 			}
 
-#if __MACOS__
-			return false;
-#else
 			return this.IsWidthConstrainedSimple() ?? (Parent as ILayoutConstraints)?.IsWidthConstrained(this) ?? false;
-#endif
 		}
 
 		bool ILayoutConstraints.IsHeightConstrained(View requester)
@@ -75,11 +70,7 @@ namespace Windows.UI.Xaml.Controls
 				return false;
 			}
 
-#if __MACOS__
-			return false;
-#else
 			return this.IsHeightConstrainedSimple() ?? (Parent as ILayoutConstraints)?.IsHeightConstrained(this) ?? false;
-#endif
 		}
 
 		public double ExtentHeight
@@ -89,7 +80,7 @@ namespace Windows.UI.Xaml.Controls
 				if (Content is FrameworkElement fe)
 				{
 					var explicitHeight = fe.Height;
-					if(!explicitHeight.IsNaN())
+					if (!explicitHeight.IsNaN())
 					{
 						return explicitHeight;
 					}
@@ -130,5 +121,65 @@ namespace Windows.UI.Xaml.Controls
 		public double ViewportHeight => DesiredSize.Height;
 
 		public double ViewportWidth => DesiredSize.Width;
+
+#if UNO_HAS_MANAGED_SCROLL_PRESENTER || __WASM__
+		protected override Size MeasureOverride(Size size)
+		{
+			if (Content is UIElement child)
+			{
+				var slotSize = size;
+
+				if (VerticalScrollBarVisibility != ScrollBarVisibility.Disabled)
+				{
+					slotSize.Height = double.PositiveInfinity;
+				}
+				if (HorizontalScrollBarVisibility != ScrollBarVisibility.Disabled)
+				{
+					slotSize.Width = double.PositiveInfinity;
+				}
+
+				child.Measure(slotSize);
+
+				return new Size(
+					Math.Min(size.Width, child.DesiredSize.Width),
+					Math.Min(size.Height, child.DesiredSize.Height)
+				);
+			}
+
+			return new Size(0, 0);
+		}
+
+		protected override Size ArrangeOverride(Size finalSize)
+		{
+			if (Content is UIElement child)
+			{
+				Rect childRect = default;
+
+				var desiredSize = child.DesiredSize;
+
+				childRect.Width = Math.Max(finalSize.Width, desiredSize.Width);
+				childRect.Height = Math.Max(finalSize.Height, desiredSize.Height);
+
+				child.Arrange(childRect);
+			}
+
+			return finalSize;
+		}
+
+		internal override bool IsViewHit()
+		{
+			return true;
+		}
+
+		void IScrollContentPresenter.OnMinZoomFactorChanged(float newValue)
+		{
+			MinimumZoomScale = newValue;
+		}
+
+		void IScrollContentPresenter.OnMaxZoomFactorChanged(float newValue)
+		{
+			MaximumZoomScale = newValue;
+		}
+#endif
 	}
 }
